requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==4.0.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
cachetools==5.3.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.6.20
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.5.14
debugpy==1.6.6
decorator==5.1.1
duckdb==0.9.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
fastapi==0.85.1
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gradio==3.40.1
gradio_client==0.7.0
greenlet==3.0.1
h11==0.14.0
hnswlib==0.7.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.1.1
ipykernel==6.23.1
ipython==8.10.0
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.3.0
kiwisolver==1.4.5
langchain==0.0.292
langsmith==0.0.63
linkify-it-py==2.0.2
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.4
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
onnxruntime==1.16.2
openai==0.27.6
orjson==3.9.10
overrides==7.4.0
packaging==23.1
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3
platformdirs==3.0.0
posthog==3.0.2
prompt-toolkit==3.0.38
protobuf==4.25.0
psutil==5.9.4
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pydantic==1.10.10
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.8.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
regex==2023.5.5
requests==2.31.0
rouge-chinese==1.0.3
rpds-py==0.12.0
scikit-learn==1.2.2
scipy==1.11.2
semantic-version==2.10.0
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.2
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.1.0
tokenizers==0.14.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
uc-micro-py==1.0.2
unstructured==0.9.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.5
websocket-client==1.5.2
websockets==11.0.3
wheel==0.41.2
yarl==1.9.2
zhipuai==1.0.7
zipp==3.11.0
zstandard==0.22.0
pymongoose==1.3.8
redis==5.1.0
python-jose==3.3.0
